% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{builder_catchments_sample}
\alias{builder_catchments_sample}
\title{Sample of catchments dataset.}
\format{
A sf object with 3263 features and 16 fields.
}
\source{
\url{https://beaconsproject.ualberta.ca/}
}
\usage{
builder_catchments_sample
}
\description{
A subset of the BEACONs catchments dataset from FDA 09EA.
}
\keyword{datasets}
