% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{existing_reserves_sample}
\alias{existing_reserves_sample}
\title{Sample of existing protected areas clipped to FDA 09EA.}
\format{
A sf object with 3 features and 2 fields.
}
\source{
\url{https://www.canada.ca/en/environment-climate-change/services/national-wildlife-areas/protected-conserved-areas-database.html}
}
\usage{
existing_reserves_sample
}
\description{
Includes three polygons making up the portion of Tombstone
Territorial Park intersecting FDA 09EA.
}
\keyword{datasets}
