% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{neighbours}
\alias{neighbours}
\title{Create a neighbours table listing neighbours for each catchment.}
\usage{
neighbours(catchments_sf)
}
\arguments{
\item{catchments_sf}{sf object of the catchments dataset with unique identifier column: CATCHNUM .}
}
\value{
A tibble of neighbouring pairs with columns \code{CATCHNUM} and \code{neighbours}.
}
\description{
For an sf object of catchments with unique CATCHNUM id's, calculates a list of neighbouring CATCHNUM pairs
and returns them in a long tibble. Neighbours are defined as having at least on point in common (within 0.1m).
}
\examples{
neighbours(catchments_sample)

}
