% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{reserve_seeds_to_builder}
\alias{reserve_seeds_to_builder}
\title{Prepares a reserve seed table for BUILDER.}
\usage{
reserve_seeds_to_builder(seeds_reserve_long)
}
\arguments{
\item{seeds_reserve_long}{Tibble or data frame output by \code{seeds_reserve()} with columns \code{Name},
\code{Areatarget} and \code{CATCHNUM}.}
}
\value{
A wide tibble with one row per reserve and CATCHNUMS separated by commas.
}
\description{
Takes a long reserve seeds table output by \code{seeds_reserve()} and converts it to the comma separated format required
by BUILDER.
}
\details{
Output of \code{reserve_seeds_to_builder} must be saved using \code{write.csv(..., row.names = FALSE, quote = FALSE)} to
be passed to BUILDER.

\code{reserve_seeds_to_builder} is called internally by \code{builder_reserve()} so is only needed if preparing a reserve seed
table to pass directly to the BUILDER software GUI (outside of R).
}
\examples{
existing_reserves_sample$areatarget_m2 <- 100000000
write.csv(
  reserve_seeds_to_builder(
    seeds_reserve(catchments_sample, existing_reserves_sample, "reserve", "areatarget_m2")),
  "reserve_seeds.csv", quote = FALSE, row.names = FALSE)
}
